Better support signed ref types when constructing PooledArray
Supports https://github.com/JuliaData/Arrow.jl/issues/113. Here's a rundown of this PR:

* PoooledArrays uses UInt32 by default as the ref type
* By default, the ref type will be widened if needed, unless the user provides themselves the ref type, in which case it will be widened and constructed, then rejected afterwards for not matching the type the user provided
* With the current code, we're then left without a way to request a signed ref type unless we know exactly what type will fit all unique elements

In this PR then, I propose removing the check that the resulting ref type matches the provided ref type, and allowing widening of signed ref types. This allows, for example, passing `PooledArray(arr, Int8)` and getting a signed ref type out, only large enough to fit the input unique elements. As for removing the error path, I content that it isn't that useful; if people are passing in a type, they're almost always passing in a type wide enough anyways, and if it comes back a little wider, is that really so bad? We could perhaps change the API more dramatically and only allow passing whether the ref types should be signed or not and allow all the widening to happen automatically.

Note that users may still always pass their own `RefArray` to ensure a specific ref type, unsigned or otherwise.